Add error for unknown RecordTypes in CSV import
QE requested validating RecordTypes during CSV import to keep user
up-to-date with amount of records which was accepted/rejected and
rejected with unknown type.

The type is processed during initial CSV
upload, so adding counter displaying user a error message with
information that there is some amount of not accepted records.

https://issues.redhat.com/browse/TACKLE-634