Introduce data migration mechanism
This adds a mechanism for automatically performing
versioned data migrations, as well as schema migrations
that GORM can't do automatically.

This also adds a migration for 2.1.0 which makes the
Application -> BusinessService relationship optional
by changing the constraint to ON DELETE SET NULL.

Updated the version of GORM to pick up a fix for a bug
in the sqlite driver that would prevent a table from
being recreated correctly.