:sparkles: Add application/facts sub-resource. (#211)
Support more granular and atomic API for managing application _facts_.

Added:
- GET /applications/facts
- GET /applications/facts/_key_.
- POST /applications/facts/_key_.
- PUT /applications/facts/_key_.  (created as needed)
- DELETE  /applications/facts/_key_.

Go binding.

- Applications.Facts()
  - List()
  - Get()
  - Set()  (created as needed)
  - Delete()
  
For backwards compat and convenience, The Application.Facts will
continue to be supported but we should recommend as _Best Practice_ for
API users to use the sub-mainly to avoid conflict.

**Note** There is a bug in the ORM where JSON fields with _number_ roots
is not scanned (unmarshalled) correctly. When fact were embedded in
Application this was not an issue. Moving each fact to a new
model/table, each fact value is now the root. Fact _number_ values
Example: Age=44 won't be supported until this is fixed in the ORM.

---------

Signed-off-by: Jeff Ortel <jortel@redhat.com>