:bug: Adjusted addon binding retry. (#281)
PROBLEM:

The addon (binding) is designed to detect and deal with temporary hub
unavailability. The primary use case is a hub pod restart. It is
intended to retry requests for 10 min. In reality, this ends up being 20
min (2 x 10). This is because in the adapter Run(), a caught error/panic
is reported to the hub which results in entering the retry logic a 2nd
time.

SOLUTION:

Maintain the error state on the client so subsequent requests will fail
immediately.

Signed-off-by: Jeff Ortel <jortel@redhat.com>