:sparkles: Add analysis total effort and include in the IssueReport. (#349)
Add analysis total effort and include in the IssueReport.

Also, noticed the IssueReport is not marshalled with `Issue` inline in
YAML. This is because the yaml:,inline tag is missing. However, the
report redefines the incidents field. This (field override) works with
the json decoder but not the yaml decoder.

Signed-off-by: Jeff Ortel <jortel@redhat.com>