:sparkles: Add sort field validation. (#383)
The approach is to build the Sort object using the fields in the model
used for the query.

Revert the usage of h.Paginated() in all but analysis.go. We can add
pagination and sorting more deliberately as needed. This is most of the
diff.

Moved the sorting object to the _new_ sort package.

Moved the BaseHandler.fields() and BaseHandler.nameOf() to a _new_
reflect package. The Fields() used for both update and inspection of
models for sort fields.

Add Sort.With() enhanced to provide validation.

---------

Signed-off-by: Jeff Ortel <jortel@redhat.com>