:sparkles: Add raw methods for error and activity reporting by addons. (#457)
Add _raw_ methods used by addons to report both errors and activity. The
reason is to reduce the number of API calls when reporting large
numbers.
For example, the _addon.command_ package reports stdout/stderr for
failed commands which can be hundreds (or thousands of lines). Calling
Activity() hundreds of times result in hundreds of API calls.
Another example is the reported analyzer errors.

Another aspect is to formalize multi-line activity. This PR codifies the
convention of:
```
- This is a list of:
- > One
- > Two
- > Three
```
and, simplifies the API.

The tackle2-addon _command_ package will need to be updated once this
merges to use RawActivity() for stdout/stderr. reporting.
The tackle2-addon-analyzer will need to updated to use both methods in
select places as well.

---------

Signed-off-by: Jeff Ortel <jortel@redhat.com>