:book: Update OpenAPI spec w/ Assessment changes (#520)
`ArchivedIssue` in the API package being an alias to another type
confuses the swag tool. To mitigate this, I had to overwrite the swagger
type of that field on the Analysis resource with `object`. This means
that the api documentation for that field will be unhelpfully vague, but
it allows the documentation to build.

Signed-off-by: Sam Lucidi <slucidi@redhat.com>