:bug: Bucket getdir (#525)
The refactor to using the new `tar` package accidentally removed setting
the X-Directory:Expand.
As a result, the client is not expanding downloaded directory (tarball).

---

Fixed API tests. The bucket directory test really needs 2+ files to be
accurate. It missed this issue.
Added the task bucket (subresource) test for symmetry.

---------

Signed-off-by: Jeff Ortel <jortel@redhat.com>