Enable importing app stakeholders
Adds two fields to the import CSV immediately before tags.
One is for the Owner and should be empty or contain a single
value in the form 'John Doe <jdoe@example.com>'. The second
is for contributors, and should be empty or a list of one or
more values in the previous format, separated by commas. As
this would be a comma-delimited field within a CSV, the whole
value should be surrounded by quotes.

Fixes https://github.com/konveyor/tackle2-hub/issues/538

Signed-off-by: Sam Lucidi <slucidi@redhat.com>