:sparkles: add addon Adapter.Log. (#550)
Add support for Adapter.Log to complement addon.Activity. This provides
convenient way for addons to include information in the pod log but not
the task activity (which is also logged). Add writers do not need to
construct their own logger.

Used in addons as:
```
addon.Log.Info("Something secret.")
```

Signed-off-by: Jeff Ortel <jortel@redhat.com>