:warning: Add "owner" and "contributors" columns to application import CSV format. (#551)
Adds two fields to the import CSV immediately before tags. One is for
the Owner and should be empty or contain a single value in the form
'John Doe <jdoe@example.com>'. The second is for contributors, and
should be empty or a list of one or more values in the previous format,
separated by commas. As this would be a comma-delimited field within a
CSV, the whole value should be surrounded by quotes. This breaks
compatibility with any existing import CSVs since it adds columns in the
middle (necessary since the number of tags is unbounded and as such the
tag list must come last).

Fixes https://github.com/konveyor/tackle2-hub/issues/538

@jortel Do you think we can get away with adding this to migration 11,
or do you think I should create a new one?

---------

Signed-off-by: Sam Lucidi <slucidi@redhat.com>