:sparkles: Merge task (kind) data. (#655)
Merge Task (kind) data into Task model data. This is to support defining
data properties in the task kind to be inherrited by models. Properties
in Task (model) data have priority.
In the interest of keeping models simple, moves TaskGroup.Propagate() to
the api. No need for it to be in the model. To support this, the Merge()
method was also moved to BaseHandler as it is used by multiple handlers.

Example:
```
---
kind: Task
apiVersion: tackle.konveyor.io/v1alpha1
metadata:
  namespace: konveyor-tackle
  name: tech-discovery
spec:
  data:
    mode:
      discovery: true
    rules:
      labels:
        included:
          - techA
          - techB
```

---------

Signed-off-by: Jeff Ortel <jortel@redhat.com>