:sparkles: Reaper optimization. (#676)
The default reaper frequency is (1) minute.

The reapers are performing 3-4 queries for every _file_ and _bucket_
evaluated for reaping. With the addition of attaching files to tasks and
task-reports, the number of files has increased substantially. The
current algorithm will run 3-4 queries for each resource evaluated.

For example: A deployment with 2000 analyzed applications can have
~25,000 files. To evaluate 4 resources for references (Task, TaskGroup,
Rule, Target) resulted in 100,000 queries each reaper run.

This PR each run (default: 1/ minute) : 
- FileReaper will run exactly **4** queries. 
- BucketReaper will run exactly **3** queries.

---------

Signed-off-by: Jeff Ortel <jortel@redhat.com>