:bug: Fix seed dup target labels (#776)
Fixes an issue with duplicate target labels.
Also, add support for a hub build version needed to detect upgrades. 
The git _tag-hash_ is written to `/etc/hub-build`. The _abbreviated_
commit hash is appended when HEAD is a commit added to the latest label.
Examples:
```
v0.6.0-beta.1-1-e161635  # tag+hash
v0.6.0-beta.1-1          # tag only
e161635                  # hash only
```

After seeded, the build is stored in settings: `.hub.db.seed.build`.
This is compared with current build to detect when the build has
changed. Seeding is _skipped_ when both the build (version) and the seed
data have not changed.

---

The go.mod update was to bump go-utils for better logging of objects.
Included here for better logging of the _Settings_.

---------

Signed-off-by: Jeff Ortel <jortel@redhat.com>