:bug: Fix error in addon: unknown user: 1001. (#794)
When updating the hub imported into the addon, loading the hub settings
produced and error: unknown user 1001. This happens when the hub setting
tries to get the current user and (I suspect) user 1001 does not exist
in addon pods.

I don't think the binding should not be loading the settings.

The addon adapter should only be loading the addon settings.

---------

Signed-off-by: Jeff Ortel <jortel@redhat.com>