:sparkles: TaskGroup modes. (#819)
Update the TaskGroup to support both batch and piipeline modes.

Mode=**Batch** (current behavior) provides a convenient method for
creating homogeneous (kind, addon, extensions) tasks. This is used by
the UI to do bulk analysis.
Mode=**Pipeline** provides a method for creating a group of
non-homogeneous tasks that are executed in order. **The driving use
case** for this is: the discovery tasks. They need to run in-order.
Further, the selection of addon and extensions for tech-discovery needs
to happen after lang-discovery so that the appropriate extension is
selected. This approach is much cleaner than the dependency rules that
causes the tech-discovery to be Postponed until lang-discovery has
completed. It solves the discovery problem because the trigger can
create and submit the tasks a a (pipeline) group. When submitted, the
manager will set only the first task state=Ready. After it completes,
the next task will be set state=Ready (and so on). Since addon and
extension selection happens when the task becomes Ready, the Language=
tag will have been set by the lang-discovery.

---------

Signed-off-by: Jeff Ortel <jortel@redhat.com>