:sparkles: Render applications using joins (#821)
Query the application using joins instead of preload() and the response
streamed back with in-line encoding.
Much faster. Reduces query of 5k applications from 10-14 seconds to 3-4
seconds. Also keeps RSS (memory) to around 100-130 MB instead of
ballooning to 6-7GB and staying there. Tested on minikube with SSD
storage.

The assessment package need to be changed so that the resolvers ALL
could be built upfront to ensure no other SQL statements were executed
while iterating the cursor (which holds the db lock).

Overlaps with #822:
- api/context.go
- api/analysis.go

After #822 is merged, these files will need to be unstaged and the PR
rebased.

Signed-off-by: Jeff Ortel <jortel@redhat.com>