:bug: fix assessment risk and confidence when listing applications. (#892)
The _resolvers_ expect the `Application.Assessments[].Sections` to be
populated.
This works for Get() because it uses db.Fetch() with
clause._Associations_ which fetches the entire nested Assessment object.
List() uses a a custom join for performance and memory footprint.

This is a regression introduced by
https://github.com/konveyor/tackle2-hub/pull/789 RE: application list
performance and memory footprint. OPTIMIZATION.

closes #891

Signed-off-by: Jeff Ortel <jortel@redhat.com>